New "View From(Q)" button shows From(Q) Data in a LIST
LIST has a FROM(Q) which the definition can be viewed on the FromQ tab.
New button opens up a Window with a VLB LIST to show all the data in the From Q.
Data can be copied to the clipboard.
List window Copy button allows copying the From Q definition.
This requires the user property {FromQ} set by .InitList().